Move user into more appropriate namespace (actually change namespace rather than just folder structure).
