+++ Sphynx/Packet/Response/MessageInfoResponsePacket.cs
@@ -1,7 +1,7 @@
 ﻿using System.Buffers;
 using System.Diagnostics.CodeAnalysis;
 using Sphynx.ChatRoom;
-using Sphynx.Core;
+using Sphynx.User;
 using Sphynx.Utils;
 
 namespace Sphynx.Packet.Response
