+++ Sphynx/Packet/Response/RegisterResponsePacket.cs
@@ -1,7 +1,7 @@
 ﻿using System.Buffers;
 using System.Diagnostics.CodeAnalysis;
 using System.Runtime.CompilerServices;
-using Sphynx.Core;
+using Sphynx.User;
 using Sphynx.Utils;
 
 namespace Sphynx.Packet.Response
